📁 [Restructure]: rate-limitモジュールをフォルダ構造に整理
- src/utils/rate-limit/ ディレクトリ構造を作成
- 機能別テストファイルに分割
- テストファイル名をfigma-url-parserと同じ命名規則に統一
- 旧rate-limit.test.tsを削除し新しい構造に移行
