+++ src/utils/index.ts
@@ -2,5 +2,5 @@ export * from './cache.js';
 export * from './case-converter.js';
 export * from './errors.js';
 export * from './logger.js';
-export * from './rate-limit.js';
+export * from './rate-limit/index.js';
 export * from './type-transformers.js';
