+++ src/utils/rate-limit/index.ts
@@ -1,0 +1,2 @@
+// Export all from rate-limit module
+export * from './rate-limit.js';
